Update monopoly_graphics.c
Added some helper functions to enable drawing with an input bitmap image. Currently this only works with input images which are 32 pixels wide and must be monochrome. Also, the file is copy pasted in hex as a char array due to inability to import files on cpulator.
Next steps include enabling variable image width and possibly using different colours. 

(currently monochrome is used because it allows for a smaller file to be hardcoded into graphics.c)