Update display.c
Added bitmap strings for letters A-Z, a smaller dollar sign and numbers 0-2 (need to add remaining numbers). Added helper function "draw_string" which takes in a string and draws it in the pixel buffer using the hardcoded bitmaps. These functions will be used to code the community and chance cards.